feat(ui): extract airdrops table element
